% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZI_ASGL.R
\name{ASGL_ZI}
\alias{ASGL_ZI}
\title{Adaptive sparse group lasso for zero inflated model}
\usage{
ASGL_ZI(
  y,
  X,
  omega = 1,
  alpha = 0.5,
  lambda = 0.01,
  family = Gamma(link = "log"),
  grouping = NULL,
  if_weight = FALSE
)
}
\arguments{
\item{y}{response matrix, matrix of size n.}

\item{X}{design matrix, matrix of size n \times p.}

\item{alpha}{scalar, balancing the l1 and l2 penalty.}

\item{lambda}{lambda.min relative to lambda.max or the lambda sequence for
the regularization path.}

\item{family}{a description of the error distribution and link function to be
used in the model. See \code{\link[stats]{family}} for details of family
functions.}

\item{grouping}{a numeric vector specifying the group of variables. Each
element of the factor/vector specifying the group of the feature.}

\item{if_weight}{if a vector of ‘prior weights’ to be used in the fitting
process. Decided based on the percentage of zero reponses.}

\item{omage}{scalar, tuning parameter for adaptive lasso}
}
\value{
A list. \describe { \item{model_fit}{a `lsgl` object with the fitted
  model.} \item{cv_fit}{a `lsgl` object with the cross validation results.}
  \item{cv_index}{the index of the best model identified by cross
  validation.} \item{cv_beta}{a vector of coefficients from the model
  selected by cv.} \item{beta_mle}{a vector of mle estimate of beta.}
  \item{alpha}{the sparse group lasso penalty}
  \item{sigmaInverseSqrt}{\hat{\Sigma}^{1/2}} \item{family}{an object of
  class "family"} \item{weight_logistic}{weight for logistic regression} }
}
\description{
Adaptive sparse group lasso for zero inflated model
}
