% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZI_ASGL.R
\name{bic_ASGL_ZI_poor_approx}
\alias{bic_ASGL_ZI_poor_approx}
\title{Calculate the Bayesian information criterion for a fitted model}
\usage{
bic_ASGL_ZI_poor_approx(fit)
}
\arguments{
\item{fit}{The fitted model from ASGL_ZI}
}
\value{

}
\description{
Calculate the Bayesian information criterion for a fitted model
}
