% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZI_ASGL.R
\name{dataSim}
\alias{dataSim}
\title{Generate simulation data}
\usage{
dataSim(
  n_obs,
  n_var,
  n_fullZero = 0,
  n_partialZero = 0,
  mean_beta = 1,
  mean_tau = 0.4,
  sd_beta = 0.7,
  sd_tau = 0.2,
  nu = 3,
  seed = 1
)
}
\arguments{
\item{n_obs}{Int, number of observations}

\item{n_var}{Int, number of variables}

\item{n_fullZero}{Int, number of variables that have no influence on the
entire model}

\item{n_partialZero}{Int, number of variables that have partial influence on
either zero or nonzero part}

\item{mean_beta}{Numeric, mean of beta for random generation, for ZERO part}

\item{mean_tau}{Numeric, mean of tau for random generation, for NONZERO part}

\item{sd_beta}{Numeric, standard deviation of beta for random generation}

\item{sd_tau}{Numeric, standard deviation of tau for random generation}

\item{nu}{Numeric, set nu (shape) for gamma distribution and sd for lognormal
distribution}

\item{seed}{Int for set.seed}
}
\value{
A list. design matrix `X`, response for ZIG `y_gamma`, response for
  ZILN `y_LN`, coefficients `beta`, `tau` , vetor of probability of y > 0
  `p_1`, binary vector indicating if y = 0 `y_1`, expectation of ZIG and ZILN
  `mu`, scalar controling the shape of gamma and lognormal `nu`=.
}
\description{
Generate simulation data
}
