% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZI_ASGL.R
\name{inverseSqrtMatrix}
\alias{inverseSqrtMatrix}
\title{Calculate the square root of the inverse of a matrix}
\usage{
inverseSqrtMatrix(x)
}
\arguments{
\item{x}{The matrix}
}
\value{
The square root of the inverse of the given matrix.
}
\description{
Use eigenvalue decomposition
}
