% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZI_ASGL.R
\name{percent_true_coef}
\alias{percent_true_coef}
\title{Calculate the the percentage of correctly identified coefficients for given betas}
\usage{
percent_true_coef(beta_est, beta_true)
}
\arguments{
\item{beta_est}{Estimated coefficients}

\item{beta_true}{True coefficients}
}
\value{
A scalalr. The percentage of correctly identified coefficients
}
\description{
Calculate the the percentage of correctly identified coefficients for given betas
}
